move packs to scripts
