+++ app/webpacker/scripts/google_analytics.js
